refactor(pipeline): split transition.rs into its five concerns
`pipeline/mod.rs` already says "Sections of the former single-file pipeline,
one per concern" and glob-re-exports each submodule. transition.rs was the
part of that split nobody finished: 2,206 lines holding five unrelated
concerns behind one name.

    transition.rs        2206 -> 765
    transition_choice.rs         396   the LLM-chosen transition
    spawn.rs                     359   ResolvedStage, setup, both spawn paths
    requirements.rs              327   children/regions/output stage-exit gates
    watchdog.rs                  312   workspace health, iteration cap, stuck
    tool_stages.rs                74   stage sync + dynamic tool refresh

`requirements.rs` rather than folding into the existing `gate.rs`: that file
is the taint gate on tool output, a different question from "may this agent
leave this stage yet".

Pure move, and the glob re-exports are what make it one: no test file, no
`world.rs` import and no call site changes, which is also why they stay. Not
one line moved between concerns; the only edit is `emit_stage_transition`
going from private to `pub(crate)`, since `enter_stage` and
`collect_transition_choice` now call it from either side of the split.

1133 runtime tests unchanged and passing.
